+++ src/model/guild/partial_guild.rs
@@ -906,7 +906,7 @@ impl PartialGuild {
         &self,
         http: impl AsRef<Http>,
         role_id: impl Into<RoleId>,
-        position: u32,
+        position: u16,
     ) -> Result<Vec<Role>> {
         self.id.edit_role_position(http, role_id, position).await
     }
